winkey daemon
I have a working daemon which will operate the winkey keyer.  It listens
to a udp port and converts the cwdaemon commands to the format winkey
needs, and sends it to the serial port the winkey unit is connected to.
After starting it goes into the background...

I have tested it with tlf as well as xtlf, and it is a good solution if
you have problems with the keying of the cwdaemon...

Version 1.0 implements the main functions:
* cw generation
* PTT lead in
* speed control
* weight control
* tune
* abort message
* a 'keyer_busy' mutex in ~/.winkey

SYNOPSIS
winkeydaemon.pl [options...]

options:
-n run in foreground for debugging
-p udp port
-d serial device
-s starting speed in wpm

One advantage is that you have a full-blown hardware keyer which is
fully integrated with your contest logger... good for rigs that lack an
internal keyer.

Next version will send the functions which are not supported (antenna
switching, rig A/B switching etc.) through to the cwdaemon socket for
handling...

I have uploaded to
http://sharon.esrac.ele.tue.nl/pub/linux/ham/tlf/winkeydaemon-1.0.tar.gz

You will probably also need to get Device::SerialPort from CPAN.

73,

Rein PA0R

Documentation is inside the script...

Retrieved from Tlf mail archive at:
	http://lists.gnu.org/archive/html/tlf-devel/2007-12/msg00009.html

Original source retrieved from Debian 5.0 (Lenny) package archive:
	http://archive.debian.org/debian/pool/main/w/winkeydaemon/winkeydaemon_1.0.1.orig.tar.gz