Add '-e' and '-m' options
This patch adds the '-e' and '-m' switches.

Using '-e' turns on 'paddle echo' on the winkeyer3.  Also, it makes the
daemon retransmit the received echo characters as UDP datagrams to the
host and port of your choice (default: localhost:6790).

Using the '-m' switch turns off the (annoying) side tone.

Tested and and found to be working with WK3SMT 3.0.9.
Works supposedly on WK2, too.

On suggestion of Wilbert, README and INSTALL have been edited to remove
some references that were no longer valid.